split out test commands update definitions into cpp
